requirements.txt
streamlit==1.45.1
google-generativeai==0.8.5
python-dotenv==1.1.0